refactor(connectLocationField): Encapsulate code to connect LocationField to store.
