+++ lib/components/form/connected-location-field.js
@@ -7,13 +7,9 @@ import {
   InputGroupAddon,
   MenuItemA
 } from '@opentripplanner/location-field/lib/styled'
-import { connect } from 'react-redux'
 import styled from 'styled-components'
 
-import { clearLocation, onLocationSelected } from '../../actions/map'
-import { addLocationSearch, getCurrentPosition } from '../../actions/location'
-import { findNearbyStops } from '../../actions/api'
-import { getActiveSearch, getShowUserSettings } from '../../util/state'
+import connectLocationField from './connect-location-field'
 
 const StyledLocationField = styled(LocationField)`
   width: 100%;
@@ -55,31 +51,4 @@ const StyledLocationField = styled(LocationField)`
   }
 `
 
-// connect to redux store
-
-const mapStateToProps = (state, ownProps) => {
-  const { config, currentQuery, location, transitIndex, user } = state.otp
-  const { currentPosition, nearbyStops, sessionSearches } = location
-  const activeSearch = getActiveSearch(state.otp)
-  const query = activeSearch ? activeSearch.query : currentQuery
-  return {
-    currentPosition,
-    geocoderConfig: config.geocoder,
-    location: query[ownProps.locationType],
-    nearbyStops,
-    sessionSearches,
-    showUserSettings: getShowUserSettings(state.otp),
-    stopsIndex: transitIndex.stops,
-    userLocationsAndRecentPlaces: [...user.locations, ...user.recentPlaces]
-  }
-}
-
-const mapDispatchToProps = {
-  addLocationSearch,
-  findNearbyStops,
-  getCurrentPosition,
-  onLocationSelected,
-  clearLocation
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(StyledLocationField)
+export default connectLocationField(StyledLocationField)
