+++ lib/components/form/intermediate-place-field.js
@@ -8,13 +8,9 @@ import {
   MenuItemA
 } from '@opentripplanner/location-field/lib/styled'
 import React, {Component} from 'react'
-import { connect } from 'react-redux'
 import styled from 'styled-components'
 
-import { clearLocation } from '../../actions/map'
-import { addLocationSearch, getCurrentPosition } from '../../actions/location'
-import { findNearbyStops } from '../../actions/api'
-import { getShowUserSettings } from '../../util/state'
+import connectLocationField from './connect-location-field'
 
 const StyledIntermediatePlace = styled(LocationField)`
   width: 100%;
@@ -78,27 +74,7 @@ class IntermediatePlaceField extends Component {
   }
 }
 
-// connect to redux store
-
-const mapStateToProps = (state, ownProps) => {
-  const { config, location, transitIndex, user } = state.otp
-  const { currentPosition, nearbyStops, sessionSearches } = location
-  return {
-    currentPosition,
-    geocoderConfig: config.geocoder,
-    nearbyStops,
-    sessionSearches,
-    showUserSettings: getShowUserSettings(state.otp),
-    stopsIndex: transitIndex.stops,
-    userLocationsAndRecentPlaces: [...user.locations, ...user.recentPlaces]
-  }
-}
-
-const mapDispatchToProps = {
-  addLocationSearch,
-  findNearbyStops,
-  getCurrentPosition,
-  clearLocation
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(IntermediatePlaceField)
+export default connectLocationField(IntermediatePlaceField, {
+  location: false,
+  onLocationSelected: false
+})
