+++ lib/components/user/favorite-location.js
@@ -1,3 +1,9 @@
+import LocationField from '@opentripplanner/location-field'
+import {
+  DropdownContainer,
+  Input as LocationFieldInput,
+  InputGroup as LocationFieldInputGroup
+} from '@opentripplanner/location-field/lib/styled'
 import React, { Component } from 'react'
 import {
   Button,
@@ -10,9 +16,36 @@ import {
 import FontAwesome from 'react-fontawesome'
 import styled from 'styled-components'
 
-import FavoriteLocationField from '../form/favorite-location-field'
+import connectLocationField from '../form/connect-location-field'
 
-// Styles
+// Style and connect LocationField to redux store.
+const StyledLocationField = styled(LocationField)`
+  width: 100%;
+  ${DropdownContainer} {
+    width: 0;
+    & > button {
+      display: none;
+    }
+  }
+  ${LocationFieldInputGroup} {
+    border: none;
+    width: 100%;
+  }
+  ${LocationFieldInput} {
+    display: table-cell;
+    font-size: 100%;
+    line-height: 20px;
+    padding: 6px 12px;
+    width: 100%;
+  }
+`
+const FavoriteLocationField = connectLocationField(StyledLocationField, {
+  clearLocation: false,
+  location: false,
+  onLocationSelected: false
+})
+
+// Styles for other controls
 const StyledDropdown = styled(DropdownButton)`
   background-color: #eee;
   width: 40px;
